add support for non-FMA MMA operators as first class SIMT citizens
This commit elevates the SIMT thread level GEMM operators as first
class citizens of the cutlass codebase. Previously the library was not
passing the Operator template parameter from the ThreadBlock level MMA
down to the SIMT MMA
from threadblock::DefaultMmaCore<>::Operator in default_mma_core_simt.h
down to gemm::warp::MmaSimt<> there was no passing of the operator
template type, and gemm::warp::MmaSimt was using arch::OpMultiplyAdd as
a default op when instantiating its thread and instruction level MMA
operation because of this.

This commit introduces a new template argument for
the SIMT MMA, sets no default for it, and adds other plumbing in
template parameters required to switch out any arch::Op with each other.

What this allows for is the reuse of the Warp and ThreadBlock level data
movement, iterators etc. while making the MMA operation itself modular. All that needs
to be done now in order to swap out MMA operation for any new semi-ring
GEMM operation as that a new Op type be specified under cutlass::arch and
implemented in the instruction level MMAs under include/cutlass/arch