add MinOp thread level epilogue operator for SemiRing Gemm
the LinearCombination operator used for BLAS MMA operations does not
suit the purpose of semiring gemm based on sum and min as the operators.

Just using the default LinearCombination operator with alpha=1 and beta=0
produces incorrect results. This is because that only stores the Semiring
product of both input matrices, srgemm(A, b), to the output matrix C.

In the event that we also want a final epilogue min of the calculated
product with the initial value of C, such as D = min(C, srgemm(A, B)),
we cannot perform that with LinearCombination epilogue.

This commit adds a new MinOp thread level epilogue to cutlass so that we
can perform a final epilogue min operation against the source value of
C matrix.

The parameter for this MinOp epilogue is a single boolean, which if set
to true at kernel invocation time, will compute this epilogue min and
store that to the output D, and if set to false, will simply store the
value of the srgemm(A, B) product to the output matrix D.