srgemm: add ability to set arbitrary init value of accumulators
this change refactors the Array::clear() method to add an overload for
an initial value and removes the previously added Array::clear_inf()
method in favour of using the new overload instead.

now the Srgemm device and kernel classes are in charge of setting this
initial value, so we surface this as an input argument in the
cutlass::gemm::device:Srgemm::Arguments struct.